refactor: move DatosPersona to Personas domain
