+++ BackendCConecta/BackendCConecta/Infraestructura/Repositorios/DatosPersona/DatosPersonaRepository.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using BackendCConecta.Aplicacion.Modulos.DatosPersona.Interfaces;
 using BackendCConecta.Infraestructura.Persistencia;
+using BackendCConecta.Dominio.Entidades.Personas;
 
 namespace BackendCConecta.Infraestructura.Repositorios.DatosPersona;
 
@@ -15,29 +16,29 @@ public class DatosPersonaRepository : IDatosPersonaRepository
 
     public async Task CrearAsync(DatosPersona entidad)
     {
-        _context.DatosPersona.Add(entidad);
+        _context.DatosPersonas.Add(entidad);
         await _context.SaveChangesAsync();
     }
 
     public async Task ActualizarAsync(DatosPersona entidad)
     {
-        _context.DatosPersona.Update(entidad);
+        _context.DatosPersonas.Update(entidad);
         await _context.SaveChangesAsync();
     }
 
     public async Task EliminarAsync(int idDatosUsuario)
     {
-        var entidad = await _context.DatosPersona.FindAsync(idDatosUsuario);
+        var entidad = await _context.DatosPersonas.FindAsync(idDatosUsuario);
         if (entidad != null)
         {
-            _context.DatosPersona.Remove(entidad);
+            _context.DatosPersonas.Remove(entidad);
             await _context.SaveChangesAsync();
         }
     }
 
     public async Task<DatosPersona?> ObtenerPorIdAsync(int idDatosUsuario)
     {
-        return await _context.DatosPersona
+        return await _context.DatosPersonas
                              .AsNoTracking()
                              .FirstOrDefaultAsync(x => x.IdDatosUsuario == idDatosUsuario);
     }
